More advanced new-game behavior
Game ends when all active players quit.
If only one player is playing he can start a new game at any
time. This exception allows the active player to include the
lobby players. Lobby players are not able to initiate the restart. This
allows the solo player to finish winning and get credit for the win.